Use relative import paths
