--- custom_components/irm_kmi/irm_kmi_api/ha.py
@@ -1,2 +1,0 @@
-
-
